coerce to unsigned long long

+++ kernel/moker/trace/syscalls.c
@@ -36,7 +36,12 @@ SYSCALL_DEFINE5(moker_sched_cbs_entity_setup,
 		int, is_hard)
 {
 	pr_info("MOKER: sys_sched_cbs_entity_setup:[pid=%d][id=%d][is_hard=%d][c=%llu][t=%llu][d=%llu]\n",
-		current->pid, id, is_hard, runtime, period, deadline);
+		current->pid,
+		id,
+		is_hard,
+		(unsigned long long)runtime,
+		(unsigned long long)period,
+		(unsigned long long)deadline);
 
 	#ifdef CONFIG_MOKER_SCHED_CBS_POLICY
 	return do_moker_sched_cbs_entity_setup(id, runtime, period, deadline, is_hard);
